refactor: componentDir -> entryDir
