+++ examples/ec-site/app/server.ts
@@ -8,7 +8,7 @@ const render = createRender()
 
 app.get('/', async (c) => {
   const products = getAllProducts()
-  const html = await render('pages/index', { products })
+  const html = await render('index', { products })
   return c.html(html)
 })
 
@@ -17,7 +17,7 @@ app.get('/products/:id', async (c) => {
   if (!product) {
     return c.text('Not found', 404)
   }
-  const html = await render('pages/detail', { product })
+  const html = await render('detail', { product })
   return c.html(html)
 })
