+++ examples/ec-site/pages/index.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
-import type { Product } from '../../app/db.ts'
-import Layout from '../Layout.vue'
+import type { Product } from '../app/db.ts'
+import Layout from '../components/Layout.vue'
 
 defineProps<{
   products: Product[]
