+++ test/integration/fixtures/pages/nested/index.vue
