+++ test/integration/fixtures/pages/static.vue
