+++ test/integration/fixtures/pages/with-css-module.vue
