+++ test/integration/fixtures/pages/with-props.vue
